% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mGOgenes.R
\name{GOgenes}
\alias{GOgenes}
\title{Mapping Genes to enriched GO terms}
\usage{
GOgenes(GOtable, GOids)
}
\arguments{
\item{GOtable}{data.frame, goana output (filtered by topGO)}

\item{GOids}{ENTREZ IDs used in goana}
}
\value{
data.frame, containing GO IDs, GOTerms, the GOtable data, and gene IDs and names.
}
\description{
Mapping Genes to enriched GO terms
}
