% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadingPlot.R
\name{loadingPlot}
\alias{loadingPlot}
\title{loadingPlot}
\usage{
loadingPlot(x, PC = "PC1", n = 10, return = FALSE)
}
\arguments{
\item{x}{Output of type prcomp. If you use mPCA, pca is a default output.}

\item{PC}{which PC to take the loadings from ('PCx')}

\item{n}{numbers of gene names to plot}

\item{return}{Boolean to determine whether to return a ggplot object or not (default = FALSE just returns the plot)}
}
\description{
Takes the rotation slot from the prcomp output object and plots the genes according to their absolute loading values.
}
