% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mFeaturePlot.R
\name{mFeaturePlot}
\alias{mFeaturePlot}
\title{Visualize 'features' on a dimensional reduction plot}
\usage{
mFeaturePlot(
  my.se,
  my.reduct = "tsne",
  my.slot = "data",
  my.features = NULL,
  colors = c("gray90", "gray90", "gray80", "yellow", "orange", "red", "darkred",
    "darkred"),
  size = 2,
  alpha = 0.4,
  order = TRUE,
  return = FALSE
)
}
\arguments{
\item{my.se}{Seurat object}

\item{my.reduct}{name of reduction to take from the embedding slot. e.g.: pca, tsne, umap ...}

\item{my.slot}{slot to take the data from. e.g.: counts, data, scale.data}

\item{my.features}{Vector of features to plot. Features must be present in modplots::gnames$Gene.name!}

\item{colors}{Vector of colors, used in scale_colour_gradientn(). Gradient from low to high expression.}

\item{size}{pt size of geom_point}

\item{alpha}{alpha parameter for geom_point}

\item{order}{Boolean determining whether to plot cells in order of expression. Can be useful if
cells expressing given feature are getting buried.}

\item{return}{Boolean, whether to return a plot (FALSE, default) or a list of ggplot objects (TRUE)}
}
\value{
A grid.arrange plot if
\code{return = FALSE}; otherwise, a list of ggplot objects
}
\description{
Colors single cells on a dimensional reduction plot according to a 'feature'
(i.e. gene expression, PC scores, number of genes detected, etc.)
}
