% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mPheatmapDESeq2.R
\name{mPheatmapDESeq2}
\alias{mPheatmapDESeq2}
\title{Custom pheatmap of DESeq2 vst data}
\usage{
mPheatmapDESeq2(vsd, IDs, ...)
}
\arguments{
\item{vsd}{vsd object (created as following: \code{vsd <- vst(dds)})}

\item{IDs}{vector of gene ids}

\item{...}{parameters passed to pheatmap}
}
\description{
Preset heatmap for DESeq2 data. Plots z scored (scaled to mean = 0 and sd = 1) vst expression of any gene list.
}
