% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mVolcanoplot.R
\name{mVolcanoplot}
\alias{mVolcanoplot}
\title{Volcano plot visualization}
\usage{
mVolcanoplot(
  table,
  type = "seurat",
  l2fc = log2(2),
  p.adj = 0.05,
  return = FALSE,
  plotly = FALSE,
  gnames = NULL
)
}
\arguments{
\item{table}{DE table created by either DESeq2 or Seurat.
For DESeq2 the table need the columns ID, log2FoldChange, and padj.
The Seurat table needs the colums gene, avg_log2_FC, and p_val_adj.}

\item{type}{character indicating the method that created the DE table. Can be either seurat, deseq2 or none. If "none" is
specified the function looks for "gene" or "id", "adj", and "2F" in the tables colnames.}

\item{l2fc}{log2 fold change threshold}

\item{p.adj}{Adjusted p-value threshold}

\item{return}{Boolean whether to return a plot (FALSE, default) or a ggplot objects (TRUE)}

\item{plotly}{Boolean whether to plot the volcano plot with plotly to create an interactive plot.}

\item{gnames}{Optional, data frame of two columns, Gene.stable.ID and Gene.name. If none is specified, internal gg6 chicken gnames table is used.}
}
\value{
A ggplot plot if \code{return = FALSE}, or else a ggplot object.
}
\description{
Intuitive way of visualizing log fold changes and adjusted pvalues
}
\details{
Adjusted pvals of 0 are capped to 1e-350 so they are not driven to infinity by -log10 transformation.
}
