% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mPCA.R
\name{mPCA}
\alias{mPCA}
\title{mPCA}
\usage{
mPCA(
  vsd,
  ntop,
  PCs = c("PC1", "PC2"),
  genes = NULL,
  meta = NULL,
  group = NULL,
  label = NULL,
  shape = NULL,
  pch = NULL,
  colors = NULL,
  pt.size = 3,
  return = FALSE
)
}
\arguments{
\item{vsd}{Variance stabilising transformed dds (vsd <- vst(dds)).}

\item{ntop}{integer, number of variable genes to use for PCA.}

\item{PCs}{character vector of length two, containing the two PCs to plot (default to: c("PC1", "PC2")).}

\item{genes}{vector of gene IDs to use instead of internally calculated variable genes.}

\item{meta}{Meta data frame for ggplot aestetics, if not specified grouping will be taken from \code{vsd`@`colData}.
If you specify a meta data frame, you have to also specify group, which should be a column of the meta data frame.}

\item{group}{Column of vsd`@`colData used as aesthetic for grouping the PCA. Meta data column if meta data is specified.}

\item{label}{Column of vsd`@`colData used as aesthetic for labeling the PCA. Meta data column if meta data is specified.}

\item{shape}{Column of vsd`@`colData used as aesthetic for point shapes in the PCA. Meta data column if meta data is specified.}

\item{pch}{vector of point characters if shape is specified, same length as factor levels in shape.}

\item{colors}{color vector, should be same length as factor levels in group}

\item{pt.size}{integer, sets pointsize in geom_point.}

\item{return}{Boolean whether to return a plot (Default = FALSE) or a ggplot object list (TRUE)}
}
\description{
Function to plot PCA biplot from a DESeq2 vst (variance stabilising transformation).
}
